+++ app/api/v1/Meta.php
@@ -8,10 +8,12 @@
 
 namespace app\api\v1;
 
+use app\exceptions\GC2Exception;
 use app\inc\Controller;
 use app\inc\Input;
 use app\inc\Route;
 use app\inc\Session;
+use app\inc\Util;
 use app\models\Layer;
 use Phpfastcache\Exceptions\PhpfastcacheInvalidArgumentException;
 use Phpfastcache\Exceptions\PhpfastcacheLogicException;
@@ -41,16 +43,13 @@ class Meta extends Controller
      * @return array
      * @throws PhpfastcacheInvalidArgumentException
      * @throws PhpfastcacheLogicException
+     * @throws GC2Exception
      */
     public function get_index()
     {
         // Get the URI params from request
         // /meta/{user}/[query]
-        $db = Route::getParam("user");
-        $dbSplit = explode("@", $db);
-        if (sizeof($dbSplit) == 2) {
-            $db = $dbSplit[1];
-        }
+        $db = Util::extractUserFromSubUserString(Route::getParam("user"))[1];
         return $this->layers->getAll($db, Session::isAuth(), Route::getParam("query"), Input::get("iex"), Input::get("parse"), Input::get("es"), false);
     }
 }