+++ app/controllers/Tilecache.php
@@ -12,6 +12,7 @@ use app\inc\Controller;
 use app\inc\Input;
 use app\conf\Connection;
 use app\conf\App;
+use app\inc\Util;
 use app\models\Database;
 use Phpfastcache\Exceptions\PhpfastcacheInvalidArgumentException;
 
@@ -33,12 +34,7 @@ class Tilecache extends Controller
     function __construct()
     {
         parent::__construct();
-
-        $this->db = Input::getPath()->part(2);
-        $dbSplit = explode("@", $this->db);
-        if (sizeof($dbSplit) == 2) {
-            $this->db = $dbSplit[1];
-        }
+        $this->db = Util::extractUserFromSubUserString(Input::getPath()->part(2))[1];
     }
 
     /**