+++ app/models/Layer.php
@@ -125,7 +125,7 @@ class Layer extends Table
      * @throws GC2Exception
      * @throws PhpfastcacheInvalidArgumentException
      */
-    public function getAll(string $db, ?bool $auth, ?string $query = null, ?bool $includeExtent = false, ?bool $parse = false, ?bool $es = false, ?bool $lookupForeignTables = true, ?array $jwt = null): array
+    public function getAll(string $db, ?bool $auth, ?string $query = null, ?bool $includeExtent = false, ?bool $parse = false, ?bool $es = false, ?bool $lookupForeignTables = true, ?array $jwt = null, bool $restriction = true): array
     {
         // If user is signed in with another user than the requested,
         // when consider the user as not signed in.
@@ -134,7 +134,7 @@ class Layer extends Table
         }
 
         $cacheType = "meta";
-        $cacheId = $this->postgisdb . "_" . Session::getUser() . "_" . $cacheType . "_" . md5($query . "_" . "(int)$auth" . "_" . (int)$includeExtent . "_" . (int)$parse . "_" . (int)$es);
+        $cacheId = $this->postgisdb . "_" . Session::getUser() . "_" . $cacheType . "_" . md5($query . "_" . "(int)$auth" . "_" . (int)$includeExtent . "_" . (int)$parse . "_" . (int)$es) . "_" . ($restriction ? 'restriction' : 'notRestriction');
 
         $CachedString = Cache::getItem($cacheId);
 
@@ -373,7 +373,7 @@ class Layer extends Table
                 } else {
                     $fieldConf = [];
                 }
-                $fields = $this->getMetaData($rel, false, true, $restrictions, null, true, $lookupForeignTables);
+                $fields = $this->getMetaData($rel, false, $restriction, $restrictions, null, true, $lookupForeignTables);
 
                 foreach ($fields as $key => $field) {
                     // If column comment is empty, we output from field conf