+++ public/js/admin/admin.js
@@ -1446,6 +1446,12 @@ $(document).ready(function () {
                                                 },
                                                 {
                                                     xtype: 'textarea',
+                                                    fieldLabel: __('QML'),
+                                                    name: 'qml',
+                                                    value: r.data.qml
+                                                },
+                                                {
+                                                    xtype: 'textarea',
                                                     height: 100,
                                                     fieldLabel: __('View definition'),
                                                     name: 'viewdefinition',
@@ -1461,13 +1467,6 @@ $(document).ready(function () {
                                                 var f = Ext.getCmp('detailform');
                                                 if (f.form.isValid()) {
                                                     var values = f.form.getValues();
-
-                                                    for (var key in values) {
-                                                        if (values.hasOwnProperty(key)) {
-                                                            values[key] = encodeURIComponent(values[key]);
-                                                        }
-                                                    }
-
                                                     var param = {
                                                         data: values
                                                     };
@@ -1485,10 +1484,8 @@ $(document).ready(function () {
                                                             store.reload();
                                                             groupsStore.load();
                                                             App.setAlert(App.STATUS_NOTICE, __("Settings updated"));
-                                                            winMoreSettings.close();
                                                         },
                                                         failure: function (response) {
-                                                            winMoreSettings.close();
                                                             Ext.MessageBox.show({
                                                                 title: 'Failure',
                                                                 msg: __(Ext.decode(response.responseText).message),
@@ -3040,6 +3037,7 @@ $(document).ready(function () {
         Ext.getCmp("a11").removeAll();
         Ext.getCmp("a12").removeAll();
         Ext.getCmp("a13").removeAll();
+        Ext.getCmp("a14").removeAll();
         Ext.getCmp("layerStylePanel").disable();
         Ext.getCmp("classTabs").disable();
 
@@ -3288,6 +3286,10 @@ $(document).ready(function () {
         a13.remove(wmsLayer.legendForm);
         a13.add(wmsLayer.legendForm);
         a13.doLayout();
+        var a14 = Ext.getCmp("a14");
+        a14.remove(wmsLayer.qmlForm);
+        a14.add(wmsLayer.qmlForm);
+        a14.doLayout();
         updateLegend();
     };
 
@@ -3767,6 +3769,20 @@ $(document).ready(function () {
                                                                     {
                                                                         xtype: "panel",
                                                                         id: "a13"
+                                                                    },
+                                                                ]
+                                                            },{
+                                                                xtype: "panel",
+                                                                title: __('QML'),
+                                                                border: false,
+                                                                defaults: {
+                                                                    border: false,
+                                                                    bodyStyle: "padding : 7px"
+                                                                },
+                                                                items: [
+                                                                    {
+                                                                        xtype: "panel",
+                                                                        id: "a14"
                                                                     },
                                                                 ]
                                                             }